requirements.txt
fastapi
gunicorn
Jinja2
pytest
uvicorn
requests
pre-commit
pydantic-settings
aiofiles
python-multipart
pillow
pytesseract
